#84 Validação feita nos schemas
