+++ backend/api/models/championships.py
@@ -4,7 +4,7 @@ from api.database.config import Base
 import enum
 from datetime import datetime
 from typing import Optional
-from pydantic import BaseModel, validator
+from pydantic import BaseModel
 
 class EnumFormat(enum.Enum):
     chaveamento = "chaveamento"
@@ -26,39 +26,6 @@ class ChampionshipInput(BaseModel):
     visibility: EnumVisibility
     admin_id: int
     game_id: int
-
-    @validator('name')
-    def name_length(cls, v):
-        if len(v) < 5 or len(v) > 100:
-            raise ValueError('O nome do campeonato deve ter entre 5 e 100 caracteres')
-        if not v.strip():
-            raise ValueError('O nome do campeonato não pode ser composto apenas por espaços em branco')
-        return v
-
-
-    @validator('start_time')
-    def start_time_not_past(cls, v):
-        if v < datetime.now():
-            raise ValueError('A data de início do campeonato não pode ser no passado')
-        return v
-
-    @validator('min_teams')
-    def min_teams_greater_than_zero(cls, v):
-        if v <= 1:
-            raise ValueError('O número mínimo de times deve ser maior que um')
-        return v
-
-    @validator('max_teams')
-    def max_teams_greater_than_min_teams(cls, v, values):
-        if 'min_teams' in values and v < values['min_teams']:
-            raise ValueError('O número máximo de times deve ser maior ou igual ao número mínimo de times')
-        return v
-
-    @validator('format')
-    def format_is_chaveamento_when_max_teams_is_not_power_of_two(cls, v, values):
-        if 'max_teams' in values and (not (values['max_teams'] & (values['max_teams'] - 1) == 0) and v != EnumFormat.chaveamento):
-            raise ValueError('O formato do campeonato deve ser chaveamento quando o número máximo de times não é potência de dois')
-        return v
 
 class ChampionshipSchema(BaseModel):
     id: int
